reorganised the integration tests
