+++ contracts/mixnet-vesting-integration-tests/src/support/mod.rs
@@ -1,5 +1,6 @@
 // Copyright 2023 - Nym Technologies SA <contact@nymtech.net>
 // SPDX-License-Identifier: Apache-2.0
 
-mod decrease_mixnode_pledge;
-mod support;
+pub mod fixtures;
+pub mod helpers;
+pub mod setup;
